+++ src/constants/routeNames.js
@@ -1,29 +1,29 @@
-const SCREEN_SUFFIX = 'Screen';
 const DRAWER_SUFFIX = 'Drawer';
-const TABBAR_SUFFIX = 'Tabbar';
 const MODAL_SUFFIX = 'Modal';
+const SCREEN_SUFFIX = 'Screen';
+const TABBAR_SUFFIX = 'Tabbar';
 
 export default {
   SCREENS: {
+    BOOKMARKS: `Bookmarks${SCREEN_SUFFIX}`,
+    BOOST: `Boost${SCREEN_SUFFIX}`,
     DRAFTS: `Drafts${SCREEN_SUFFIX}`,
     EDITOR: `Editor${SCREEN_SUFFIX}`,
     FOLLOWS: `Follows${SCREEN_SUFFIX}`,
+    FREE_ESTM: `FreeEstm${SCREEN_SUFFIX}`,
     HOME: `Home${SCREEN_SUFFIX}`,
     LOGIN: `Login${SCREEN_SUFFIX}`,
     PINCODE: `PinCode${SCREEN_SUFFIX}`,
     POST: `Post${SCREEN_SUFFIX}`,
     PROFILE: `Profile${SCREEN_SUFFIX}`,
+    PROMOTE: `Promote${SCREEN_SUFFIX}`,
+    REBLOGS: `Reblogs${SCREEN_SUFFIX}`,
+    SEARCH_RESULT: `SearchResult${SCREEN_SUFFIX}`,
     SETTINGS: `Settings${SCREEN_SUFFIX}`,
     STEEM_CONNECT: `SteemConnect${SCREEN_SUFFIX}`,
-    VOTERS: `Voters${SCREEN_SUFFIX}`,
-    BOOKMARKS: `Bookmarks${SCREEN_SUFFIX}`,
-    SEARCH_RESULT: `SearchResult${SCREEN_SUFFIX}`,
     TRANSFER: `Transfer${SCREEN_SUFFIX}`,
-    BOOST: `Boost${SCREEN_SUFFIX}`,
-    BOOST_POST: `BoostPost${SCREEN_SUFFIX}`,
-    PROMOTE: `Promote${SCREEN_SUFFIX}`,
-    FREE_ESTM: `FreeEstm${SCREEN_SUFFIX}`,
-    REBLOGS: `Reblogs${SCREEN_SUFFIX}`,
+    VOTERS: `Voters${SCREEN_SUFFIX}`,
+    REDEEM: `Redeem${SCREEN_SUFFIX}`,
   },
   DRAWER: {
     MAIN: `Main${DRAWER_SUFFIX}`,
@@ -31,8 +31,8 @@ export default {
   TABBAR: {
     HOME: `Home${TABBAR_SUFFIX}`,
     NOTIFICATION: `Notification${TABBAR_SUFFIX}`,
-    POSTBUTTON: `PostButton${TABBAR_SUFFIX}`,
     POINTS: `Points${TABBAR_SUFFIX}`,
+    POSTBUTTON: `PostButton${TABBAR_SUFFIX}`,
     PROFILE: `Profile${TABBAR_SUFFIX}`,
   },
   MODAL: {
