+++ src/containers/index.js
@@ -1,4 +1,5 @@
 import PointsContainer from './pointsContainer';
 import TransferContainer from './transferContainer';
+import RedeemContainer from './redeemContainer';
 
-export { PointsContainer, TransferContainer };
+export { PointsContainer, TransferContainer, RedeemContainer };
