+++ src/containers/pointsContainer.js
@@ -8,7 +8,7 @@ import { withNavigation } from 'react-navigation';
 // Services and Actions
 import { getUser, getUserPoints, claim } from '../providers/esteem/ePoint';
 import { openPinCodeModal } from '../redux/actions/applicationActions';
-import { promote, getAccount, boost } from '../providers/steem/dsteem';
+import { getAccount, boost } from '../providers/steem/dsteem';
 import { getUserDataWithUsername } from '../realm/realm';
 import { toastNotification } from '../redux/actions/uiAction';
 
@@ -89,16 +89,18 @@ class PointsContainer extends Component {
         break;
 
       case 1:
-        navigateTo = ROUTES.SCREENS.PROMOTE;
+        navigateTo = ROUTES.SCREENS.REDEEM;
         navigateParams = {
           balance,
+          redeemType: 'promote',
         };
         break;
 
       case 2:
-        navigateTo = ROUTES.SCREENS.BOOST_POST;
+        navigateTo = ROUTES.SCREENS.REDEEM;
         navigateParams = {
           balance,
+          redeemType: 'boost',
         };
         break;
 
@@ -194,28 +196,6 @@ class PointsContainer extends Component {
     this.setState({ isClaiming: false });
   };
 
-  _promote = async (duration, permlink, author, user) => {
-    const { currentAccount, pinCode, dispatch, intl, navigation } = this.props;
-    this.setState({ isLoading: true });
-
-    await promote(user || currentAccount, pinCode, duration, permlink, author)
-      .then(() => {
-        this.setState({ isLoading: false });
-        navigation.goBack();
-        dispatch(toastNotification(intl.formatMessage({ id: 'alert.successful' })));
-      })
-      .catch(error => {
-        if (error) {
-          Alert.alert(
-            `Fetching data from server failed, please try again or notify us at info@esteem.app \n${error.message.substr(
-              0,
-              20,
-            )}`,
-          );
-        }
-      });
-  };
-
   _boost = async (point, permlink, author, user) => {
     const { currentAccount, pinCode, dispatch, intl, navigation } = this.props;
     this.setState({ isLoading: true });
@@ -274,10 +254,10 @@ class PointsContainer extends Component {
         isDarkTheme,
         isLoading,
         navigationParams,
-        promote: this._promote,
         refreshing,
         userActivities,
         userPoints,
+        redeemType: get(navigationParams, 'redeemType'),
       })
     );
   }
