+++ src/navigation/routes.js
@@ -9,6 +9,8 @@ import { default as ROUTES } from '../constants/routeNames';
 // Screens
 import {
   Bookmarks,
+  Boost,
+  BoostPost,
   Drafts,
   Editor,
   Follows,
@@ -16,15 +18,13 @@ import {
   PinCode,
   Post,
   Profile,
+  Reblogs,
+  Redeem,
+  SearchResult,
   Settings,
   SteemConnect,
-  Voters,
-  SearchResult,
   Transfer,
-  Boost,
-  Promote,
-  BoostPost,
-  Reblogs,
+  Voters,
 } from '../screens';
 
 // Components
@@ -115,14 +115,8 @@ const stackNavigatior = createStackNavigator(
         header: () => null,
       },
     },
-    [ROUTES.SCREENS.PROMOTE]: {
-      screen: Promote,
-      navigationOptions: {
-        header: () => null,
-      },
-    },
-    [ROUTES.SCREENS.BOOST_POST]: {
-      screen: BoostPost,
+    [ROUTES.SCREENS.REDEEM]: {
+      screen: Redeem,
       navigationOptions: {
         header: () => null,
       },
