+++ src/screens/index.js
@@ -15,10 +15,10 @@ import { SearchResult } from './searchResult';
 import { Settings } from './settings';
 import Voters from './voters';
 import BoostPost from './boostPost/screen/boostPostScreen';
-import Promote from './promote/screen/promoteScreen';
 import SteemConnect from './steem-connect/steemConnect';
 import Transfer from './transfer';
 import Reblogs from './reblogs';
+import Redeem from './redeem/screen/redeemScreen';
 
 export {
   Bookmarks,
@@ -35,11 +35,11 @@ export {
   Points,
   Post,
   Profile,
-  Promote,
   SearchResult,
   Settings,
   SteemConnect,
   Transfer,
   Voters,
   Reblogs,
+  Redeem,
 };
